feat(user): support sso and api key
