+++ pkg/modules/user/user.go
@@ -3,10 +3,12 @@ package user
 import (
 	"context"
 	"net/http"
+	"net/url"
 
 	"github.com/SigNoz/signoz/pkg/types"
 	"github.com/SigNoz/signoz/pkg/types/authtypes"
 	"github.com/SigNoz/signoz/pkg/valuer"
+	"github.com/google/uuid"
 )
 
 type Module interface {
@@ -47,6 +49,12 @@ type Module interface {
 
 	// Auth Domain
 	GetAuthDomainByEmail(ctx context.Context, email string) (*types.GettableOrgDomain, error)
+	GetDomainFromSsoResponse(ctx context.Context, url *url.URL) (*types.GettableOrgDomain, error)
+
+	ListDomains(ctx context.Context, orgID valuer.UUID) ([]*types.GettableOrgDomain, error)
+	CreateDomain(ctx context.Context, domain *types.GettableOrgDomain) error
+	UpdateDomain(ctx context.Context, doamin *types.GettableOrgDomain) error
+	DeleteDomain(ctx context.Context, id uuid.UUID) error
 
 	// API KEY
 	CreateAPIKey(ctx context.Context, apiKey *types.StorableAPIKey) error
@@ -85,4 +93,9 @@ type Handler interface {
 	ListAPIKeys(http.ResponseWriter, *http.Request)
 	UpdateAPIKey(http.ResponseWriter, *http.Request)
 	RevokeAPIKey(http.ResponseWriter, *http.Request)
+
+	ListDomains(http.ResponseWriter, *http.Request)
+	CreateDomain(http.ResponseWriter, *http.Request)
+	UpdateDomain(http.ResponseWriter, *http.Request)
+	DeleteDomain(http.ResponseWriter, *http.Request)
 }
